chore(discover-quick-context): Directory structure.
