+++ static/app/views/eventsV2/table/quickContext/releaseContext.spec.tsx
@@ -1,0 +1,96 @@
+import {QueryClient, QueryClientProvider} from '@tanstack/react-query';
+
+import {render, screen, within} from 'sentry-test/reactTestingLibrary';
+
+import ConfigStore from 'sentry/stores/configStore';
+
+import ReleaseContext from './releaseContext';
+import {defaultRow, mockedReleaseWithHealth, mockedUser1} from './testUtils';
+
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      retry: false,
+    },
+  },
+});
+
+const renderReleaseContext = () => {
+  const organization = TestStubs.Organization();
+  render(
+    <QueryClientProvider client={queryClient}>
+      <ReleaseContext dataRow={defaultRow} organization={organization} />
+    </QueryClientProvider>,
+    {organization}
+  );
+};
+
+describe('Quick Context Content Release Column', function () {
+  afterEach(() => {
+    queryClient.clear();
+    MockApiClient.clearMockResponses();
+  });
+
+  it('Renders Release details for release', async () => {
+    MockApiClient.addMockResponse({
+      url: '/organizations/org-slug/releases/backend@22.10.0+aaf33944f93dc8fa4234ca046a8d88fb1dccfb76/',
+      body: mockedReleaseWithHealth,
+    });
+
+    renderReleaseContext();
+
+    expect(await screen.findByText(/Created/i)).toBeInTheDocument();
+    expect(screen.getByText(/7 years ago/i)).toBeInTheDocument();
+    expect(screen.getByText(/Last Event/i)).toBeInTheDocument();
+    expect(screen.getByText(/6 years ago/i)).toBeInTheDocument();
+    expect(screen.getByText(/New Issues/i)).toBeInTheDocument();
+    expect(screen.getByText(/21/i)).toBeInTheDocument();
+  });
+
+  it('Renders Last Commit', async () => {
+    MockApiClient.addMockResponse({
+      url: '/organizations/org-slug/releases/backend@22.10.0+aaf33944f93dc8fa4234ca046a8d88fb1dccfb76/',
+      body: mockedReleaseWithHealth,
+    });
+
+    renderReleaseContext();
+
+    expect(await screen.findByText(/Last Commit/i)).toBeInTheDocument();
+    expect(screen.getByTestId('quick-context-commit-row')).toBeInTheDocument();
+  });
+
+  it('Renders Commit Count and Author when user is NOT in list of authors', async () => {
+    MockApiClient.addMockResponse({
+      url: '/organizations/org-slug/releases/backend@22.10.0+aaf33944f93dc8fa4234ca046a8d88fb1dccfb76/',
+      body: mockedReleaseWithHealth,
+    });
+
+    renderReleaseContext();
+
+    const authorsSectionHeader = within(
+      await screen.findByTestId('quick-context-release-author-header')
+    );
+
+    expect(authorsSectionHeader.getByText(/4/i)).toBeInTheDocument();
+    expect(authorsSectionHeader.getByText(/commits by/i)).toBeInTheDocument();
+    expect(authorsSectionHeader.getByText(/2/i)).toBeInTheDocument();
+    expect(authorsSectionHeader.getByText(/authors/i)).toBeInTheDocument();
+    expect(screen.getByText(/KN/i)).toBeInTheDocument();
+    expect(screen.getByText(/VN/i)).toBeInTheDocument();
+  });
+
+  it('Renders Commit Count and Author when user is in list of authors', async () => {
+    jest.spyOn(ConfigStore, 'get').mockImplementation(() => mockedUser1);
+    MockApiClient.addMockResponse({
+      url: '/organizations/org-slug/releases/backend@22.10.0+aaf33944f93dc8fa4234ca046a8d88fb1dccfb76/',
+      body: mockedReleaseWithHealth,
+    });
+
+    renderReleaseContext();
+
+    expect(await screen.findByText(/4/i)).toBeInTheDocument();
+    expect(screen.getByText(/commits by you and 1 other/i)).toBeInTheDocument();
+    expect(screen.getByText(/KN/i)).toBeInTheDocument();
+    expect(screen.getByText(/VN/i)).toBeInTheDocument();
+  });
+});
